+++ tests/core_test.py
@@ -26,7 +26,7 @@ from jax import core
 from jax import numpy as jnp
 from jax import test_util as jtu
 from jax.api import jvp, linearize, vjp, jit
-from jax.lax import UnshapedArray, ShapedArray, ConcreteArray
+from jax.core import UnshapedArray, ShapedArray, ConcreteArray
 from jax.tree_util import tree_flatten, tree_unflatten, tree_multimap, tree_reduce, tree_leaves
 from jax.util import partial
 from jax.interpreters import partial_eval as pe